clean code structure for models
