+++ django_project/core/settings/base.py
@@ -4,9 +4,8 @@ core.settings.base
 """
 # Django settings for projecta project.
 
-from django.contrib.messages import constants as messages
-
 from .utils import ABS_PATH
+from django.contrib.messages import constants as messages
 
 ADMINS = (
     ('Tim Sutton', 'tim@kartoza.com'),
@@ -45,7 +44,7 @@ USE_TZ = True
 
 # Absolute filesystem path to the directory that will hold user-uploaded files.
 # Example: "/var/www/example.com/media/"
-MEDIA_ROOT = ABS_PATH('media')
+MEDIA_ROOT = '/home/web/media'
 
 # URL that handles the media served from MEDIA_ROOT. Make sure to use a
 # trailing slash.
@@ -58,7 +57,7 @@ MEDIA_URL = '/media/'
 # Don't put anything in this directory yourself; store your static files
 # in apps' "static/" subdirectories and in STATICFILES_DIRS.
 # Example: "/var/www/example.com/static/"
-STATIC_ROOT = ABS_PATH('static')
+STATIC_ROOT = '/home/web/static'
 
 # URL prefix for static files.
 # Example: "http://example.com/static/", "http://static.example.com/"
@@ -82,7 +81,7 @@ STATICFILES_FINDERS = (
 
 # import SECRET_KEY into current namespace
 # noinspection PyUnresolvedReferences
-from .secret import SECRET_KEY  # noqa # isort:skip
+from .secret import SECRET_KEY  # noqa
 
 # List of callables that know how to import templates from various sources.
 TEMPLATE_LOADERS = (
@@ -168,5 +167,5 @@ MESSAGE_TAGS = {
     messages.INFO: 'info',
     messages.SUCCESS: 'success',
     messages.WARNING: 'warning',
-    messages.ERROR: 'danger'  # 'error' by default
+    messages.ERROR: 'danger' # 'error' by default
 }
