+++ django_project/localities/models/__init__.py
@@ -1,0 +1,12 @@
+__author__ = 'Irwan Fathurrahman <irwan@kartoza.com>'
+__date__ = '07/01/19'
+
+from localities.models.attribute import *
+from localities.models.boundary import *
+from localities.models.changeset import *
+from localities.models.data_loader import *
+from localities.models.domain import *
+from localities.models.locality import *
+from localities.models.masterization import *
+from localities.models.specification import *
+from localities.models.value import *
